# design_docs/openapi.v0.yaml
openapi: 3.0.3
info:
  title: Curious Now API (v0)
  version: "0.1"
servers:
  - url: /v1
paths:
  /items/feed:
    get:
      summary: Stage 1 Item feed (debug/internal after Stage 2)
      parameters:
        - in: query
          name: page
          schema: { type: integer, minimum: 1, default: 1 }
        - in: query
          name: page_size
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
        - in: query
          name: source_id
          schema: { type: string, format: uuid }
        - in: query
          name: content_type
          schema: { $ref: "#/components/schemas/ContentType" }
      responses:
        "200":
          description: Paginated list of Items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemsFeedResponse"
...
  /entities:
    get:
      summary: Stage 10 entities list/search
      parameters:
        - in: query
          name: q
          required: false
          schema: { type: string }
        - in: query
          name: entity_type
          required: false
          schema: { $ref: "#/components/schemas/EntityType" }
        - in: query
          name: page
          schema: { type: integer, minimum: 1, default: 1 }
        - in: query
          name: page_size
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      responses:
        "200":
          description: Entities
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EntitiesResponse" }
  /entities/{id}:
    get:
      summary: Stage 10 entity detail
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Entity detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EntityDetail" }
        "301":
          description: Entity has been merged; client should redirect to canonical entity id
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EntityRedirectResponse" }
  /user/follows/entities:
    get:
      summary: Stage 10 list followed entities (convenience)
      security:
        - SessionAuth: []
      responses:
        "200":
          description: Followed entities
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserFollowedEntitiesResponse" }
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /user/follows/entities/{entity_id}:
    post:
      summary: Stage 10 follow an entity
      security:
        - SessionAuth: []
      parameters:
        - in: path
          name: entity_id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Followed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SimpleOkResponse" }
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
    delete:
      summary: Stage 10 unfollow an entity
      security:
        - SessionAuth: []
      parameters:
        - in: path
          name: entity_id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Unfollowed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SimpleOkResponse" }
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  /admin/entities:
    post:
      summary: Stage 10 create an entity
      security:
        - AdminTokenAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminEntityCreateRequest" }
      responses:
        "200":
          description: Created entity
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Entity" }
  /admin/entities/{id}:
    patch:
      summary: Stage 10 update an entity
      security:
        - AdminTokenAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminEntityPatchRequest" }
      responses:
        "200":
          description: Updated entity
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Entity" }
  /admin/entities/{id}/merge:
    post:
      summary: Stage 10 merge an entity into another (creates redirect)
      security:
        - AdminTokenAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminEntityMergeRequest" }
      responses:
        "200":
          description: Merge completed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminEntityMergeResponse" }
  /admin/clusters/{id}/entities:
    put:
      summary: Stage 10 set cluster entities (manual + locked)
      security:
        - AdminTokenAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminSetClusterEntitiesRequest" }
      responses:
        "200":
          description: Updated cluster entities
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SimpleOkResponse" }
  /admin/experiments:
    post:
      summary: Stage 10 create an experiment
      security:
        - AdminTokenAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminExperimentCreateRequest" }
      responses:
        "200":
          description: Created experiment
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Experiment" }
  /admin/experiments/{id}:
    patch:
      summary: Stage 10 update an experiment
      security:
        - AdminTokenAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminExperimentPatchRequest" }
      responses:
        "200":
          description: Updated experiment
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Experiment" }
  /admin/feature_flags/{key}:
    put:
      summary: Stage 10 upsert a feature flag
      security:
        - AdminTokenAuth: []
      parameters:
        - in: path
          name: key
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminFeatureFlagUpsertRequest" }
      responses:
        "200":
          description: Updated flag
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FeatureFlag" }
  /admin/source_pack/import:
    post:
      summary: Admin import of a source pack (idempotent upsert)
      security:
        - AdminTokenAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SourcePack" }
      responses:
        "200":
          description: Import summary
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SourcePackImportResponse" }
  /admin/sources/{id}:
    patch:
      summary: Admin update a Source (active toggle, metadata)
      security:
        - AdminTokenAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminSourcePatch" }
      responses:
        "200":
          description: Updated source
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Source" }
  /admin/feeds/{id}:
    patch:
      summary: Admin update a SourceFeed (active toggle, fetch interval)
      security:
        - AdminTokenAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminFeedPatch" }
      responses:
        "200":
          description: Updated feed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SourceFeedHealth" }
  /admin/ingestion/run:
    post:
      summary: Admin trigger a one-off ingestion run (best-effort)
      security:
        - AdminTokenAuth: []
      parameters:
        - in: query
          name: feed_id
          required: false
          schema: { type: string, format: uuid }
      responses:
        "202":
          description: Accepted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminRunResponse" }
components:
  securitySchemes:
    AdminTokenAuth:
      type: apiKey
      in: header
      name: X-Admin-Token
    SessionAuth:
      type: apiKey
      in: cookie
      name: cn_session
  schemas:
    ContentType:
      type: string
      enum: [news, press_release, preprint, peer_reviewed, report]
    EntityType:
      type: string
      enum: [person, institution, model, dataset, method, venue]
    Source:
      type: object
      required: [source_id, name, source_type, active]
      properties:
        source_id: { type: string, format: uuid }
        name: { type: string }
        homepage_url: { type: string, nullable: true }
        source_type:
          type: string
          enum: [journalism, journal, preprint_server, university, government, lab, blog]
        reliability_tier:
          type: string
          nullable: true
          enum: [tier1, tier2, tier3]
        active: { type: boolean }
    SourceFeedHealth:
      type: object
      required: [feed_id, feed_url, feed_type, active]
      properties:
        feed_id: { type: string, format: uuid }
        feed_url: { type: string }
        feed_type: { type: string, enum: [rss, atom, api] }
        active: { type: boolean }
        fetch_interval_minutes: { type: integer }
        last_fetched_at: { type: string, format: date-time, nullable: true }
        last_status: { type: integer, nullable: true }
        error_streak: { type: integer }
...
    Item:
      type: object
      required: [item_id, canonical_url, title, source, content_type, fetched_at]
      properties:
        item_id: { type: string, format: uuid }
        canonical_url: { type: string }
        url: { type: string }
        title: { type: string }
        snippet: { type: string, nullable: true }
        author: { type: string, nullable: true }
        published_at: { type: string, format: date-time, nullable: true }
        fetched_at: { type: string, format: date-time }
        content_type: { $ref: "#/components/schemas/ContentType" }
        paywalled: { type: boolean, nullable: true }
        source:
          type: object
          required: [source_id, name]
          properties:
            source_id: { type: string, format: uuid }
            name: { type: string }
        arxiv_id: { type: string, nullable: true }
        doi: { type: string, nullable: true }
        pmid: { type: string, nullable: true }
    ItemsFeedResponse:
      type: object
      required: [page, results]
      properties:
        page: { type: integer }
        results:
          type: array
          items: { $ref: "#/components/schemas/Item" }
    TopicChip:
      type: object
      required: [topic_id, name, score]
      properties:
        topic_id: { type: string, format: uuid }
        name: { type: string }
        score: { type: number }
    ClusterCard:
      type: object
      required: [cluster_id, canonical_title, updated_at, distinct_source_count]
      properties:
        cluster_id: { type: string, format: uuid }
        canonical_title: { type: string }
        updated_at: { type: string, format: date-time }
        distinct_source_count: { type: integer }
        top_topics:
          type: array
          items: { $ref: "#/components/schemas/TopicChip" }
        content_type_badges:
          type: array
          items: { $ref: "#/components/schemas/ContentType" }
        method_badges:
          type: array
          items: { type: string }
        takeaway: { type: string, nullable: true }
        confidence_band:
          type: string
          nullable: true
          enum: [early, growing, established]
        anti_hype_flags:
          type: array
          items: { type: string }
        is_saved: { type: boolean, nullable: true }
        is_watched: { type: boolean, nullable: true }
        why_in_feed: { $ref: "#/components/schemas/WhyInFeed" }
...
    EntityRedirectResponse:
      type: object
      required: [redirect_to_entity_id]
      properties:
        redirect_to_entity_id: { type: string, format: uuid }
...
    SourcePack:
      type: object
      required: [sources]
      properties:
        sources:
          type: array
          items:
            type: object
            required: [name, source_type, feeds]
            properties:
              name: { type: string }
              homepage_url: { type: string, nullable: true }
              source_type:
                type: string
                enum: [journalism, journal, preprint_server, university, government, lab, blog]
              reliability_tier:
                type: string
                nullable: true
                enum: [tier1, tier2, tier3]
              active: { type: boolean, nullable: true }
              terms_notes: { type: string, nullable: true }
              feeds:
                type: array
                items:
                  type: object
                  required: [feed_url, feed_type]
                  properties:
                    feed_url: { type: string }
                    feed_type: { type: string, enum: [rss, atom, api] }
                    fetch_interval_minutes: { type: integer, nullable: true }
                    active: { type: boolean, nullable: true }
    SourcePackImportResponse:
      type: object
      required: [sources_upserted, feeds_upserted]
      properties:
        sources_upserted: { type: integer }
        feeds_upserted: { type: integer }
    AdminSourcePatch:
      type: object
      properties:
        name: { type: string }
        homepage_url: { type: string, nullable: true }
        source_type:
          type: string
          enum: [journalism, journal, preprint_server, university, government, lab, blog]
        reliability_tier:
          type: string
          nullable: true
          enum: [tier1, tier2, tier3]
        terms_notes: { type: string, nullable: true }
        active: { type: boolean }
    AdminFeedPatch:
      type: object
      properties:
        feed_url: { type: string }
        feed_type: { type: string, enum: [rss, atom, api] }
        fetch_interval_minutes: { type: integer }
        active: { type: boolean }
    AdminRunResponse:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [accepted] }
    WhyInFeed:
      type: object
      properties:
        reasons:
          type: array
          items: { type: string }
        matched_topic_ids:
          type: array
          items: { type: string, format: uuid }
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
    SimpleOkResponse:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [ok] }
...
    Entity:
      type: object
      required: [entity_id, entity_type, name]
      properties:
        entity_id: { type: string, format: uuid }
        entity_type: { $ref: "#/components/schemas/EntityType" }
        name: { type: string }
        description_short: { type: string, nullable: true }
        external_url: { type: string, nullable: true }
    EntityChip:
      allOf:
        - $ref: "#/components/schemas/Entity"
        - type: object
          properties:
            is_followed: { type: boolean, nullable: true }
    EntityDetail:
      allOf:
        - $ref: "#/components/schemas/EntityChip"
        - type: object
          properties:
            latest_clusters:
              type: array
              items: { $ref: "#/components/schemas/ClusterCard" }
            related_entities:
              type: array
              items: { $ref: "#/components/schemas/EntityChip" }
    EntitiesResponse:
      type: object
      required: [page, results]
      properties:
        page: { type: integer }
        results:
          type: array
          items: { $ref: "#/components/schemas/EntityChip" }
    UserFollowedEntitiesResponse:
      type: object
      required: [entities]
      properties:
        entities:
          type: array
          items: { $ref: "#/components/schemas/EntityChip" }
    AdminEntityCreateRequest:
      type: object
      required: [entity_type, name]
      properties:
        entity_type: { $ref: "#/components/schemas/EntityType" }
        name: { type: string }
        description_short: { type: string, nullable: true }
        external_url: { type: string, nullable: true }
    AdminEntityPatchRequest:
      type: object
      properties:
        entity_type: { $ref: "#/components/schemas/EntityType" }
        name: { type: string }
        description_short: { type: string, nullable: true }
        external_url: { type: string, nullable: true }
    AdminEntityMergeRequest:
      type: object
      required: [to_entity_id]
      properties:
        to_entity_id: { type: string, format: uuid }
        notes: { type: string, nullable: true }
        supporting_item_ids:
          type: array
          items: { type: string, format: uuid }
    AdminEntityMergeResponse:
      type: object
      required: [from_entity_id, to_entity_id]
      properties:
        from_entity_id: { type: string, format: uuid }
        to_entity_id: { type: string, format: uuid }
    AdminSetClusterEntitiesRequest:
      type: object
      required: [entities]
      properties:
        replace:
          type: boolean
          default: true
        entities:
          type: array
          items: { $ref: "#/components/schemas/AdminClusterEntityAssignment" }
        notes: { type: string, nullable: true }
    AdminClusterEntityAssignment:
      type: object
      required: [entity_id]
      properties:
        entity_id: { type: string, format: uuid }
        score: { type: number, nullable: true }
        locked: { type: boolean, nullable: true }
    Experiment:
      type: object
      required: [experiment_id, key, active]
      properties:
        experiment_id: { type: string, format: uuid }
        key: { type: string }
        description: { type: string, nullable: true }
        active: { type: boolean }
        start_at: { type: string, format: date-time, nullable: true }
        end_at: { type: string, format: date-time, nullable: true }
    AdminExperimentCreateRequest:
      type: object
      required: [key]
      properties:
        key: { type: string }
        description: { type: string, nullable: true }
        active: { type: boolean, nullable: true }
        start_at: { type: string, format: date-time, nullable: true }
        end_at: { type: string, format: date-time, nullable: true }
    AdminExperimentPatchRequest:
      type: object
      properties:
        description: { type: string, nullable: true }
        active: { type: boolean }
        start_at: { type: string, format: date-time, nullable: true }
        end_at: { type: string, format: date-time, nullable: true }
    FeatureFlag:
      type: object
      required: [key, enabled]
      properties:
        key: { type: string }
        enabled: { type: boolean }
        config:
          type: object
          additionalProperties: true
    AdminFeatureFlagUpsertRequest:
      type: object
      properties:
        enabled: { type: boolean }
        config:
          type: object
          additionalProperties: true
...